+++ public/assets/css/main.css
@@ -93,13 +93,12 @@ a:active {
 }
 
 /* RESPONSIVE TABLET */
-@media (min-width: 1366px) {
+@media (min-width: 1007px) {
     footer {
         background-color: black;
         position: absolute;
         bottom: 0;
         width: 100%;
-        /* height: 40px; */
         color: white;
     }
 
@@ -107,12 +106,12 @@ a:active {
         height: 15rem;
     }
 
-    html,
+    /* html,
     body {
         height: 100%;
         margin: 0;
         padding: 0;
         display: flex;
         flex-direction: column;
-    }
+    } */
 }
